hud: placeholder command to start hud
