+++ internal/cli/cli.go
@@ -45,6 +45,7 @@ func Execute() {
 	}
 
 	addCommand(rootCmd, &upCmd{browserMode: engine.BrowserAuto})
+	addCommand(rootCmd, &hudCmd{})
 	rootCmd.PersistentFlags().BoolVarP(&debug, "debug", "d", false, "Enable debug logging")
 	rootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "Enable verbose logging")
 
@@ -125,4 +126,23 @@ func addCommand(parent *cobra.Command, child tiltCmd) {
 	}
 
 	parent.AddCommand(cobraChild)
+}
+
+// Returns a build datestamp in the format 2018-08-30
+func buildDateStamp() string {
+	// TODO(nick): Add a mechanism to encode the datestamp in the binary with
+	// ldflags. This currently only works if you are building your own
+	// binaries. It won't work once we're distributing pre-built binaries.
+	path, err := os.Executable()
+	if err != nil {
+		return "[unknown]"
+	}
+
+	info, err := os.Stat(path)
+	if err != nil {
+		return "[unknown]"
+	}
+
+	modTime := info.ModTime()
+	return modTime.Format("2006-01-02")
 }
